+++ src/doppler.cpp
@@ -280,281 +280,3 @@ double get_doppler_centroid_rate(PACKET_VEC_1D& burst_packets, const double& vel
 
     return (-2 * velocity * antenna_steering_rate) / WAVELENGTH;
 }
-
-
-// Mosaic a Signal for UFR
-CF_VEC_1D get_tiled_signal(CF_VEC_1D& signal, const double& num_replicas)
-{
-    int num_az = signal.size();
-    int shape = std::floor(num_replicas * num_az);
-    int offset = std::floor(num_replicas);
-
-    CF_VEC_1D tiled_signal(shape);
-
-    // Mosaic `floor(num_replicas)` times
-    for (int i = 0; i < offset; i++)
-    {
-        std::transform(
-            signal.begin(), signal.end(),
-                tiled_signal.begin() + i * num_az,
-                    [] (std::complex<double>& sample) {
-                        return sample;
-                    }
-        );
-    }
-
-    // If `num_replicas` isn't a whole number, tile the remaining percentage after `floor(num_replicas)`
-    if (std::floor(num_replicas) != num_replicas)
-    {
-        int num_samples = std::floor((num_replicas - std::floor(num_replicas)) * num_az);
-
-        std::transform(
-            signal.begin(), signal.begin() + num_samples,
-                tiled_signal.begin() + offset * num_az,
-                    [] (std::complex<double>& sample) {
-                        return sample;
-                    }
-        );
-    }
-
-    return tiled_signal;
-}
-
-
-CF_VEC_2D azimuth_frequency_ufr(
-    CF_VEC_2D& range_compressed,
-    F_VEC_1D&  dc_estimates,
-    L0Packet&  initial_packet,
-    const double& dc_rate,
-    const double& burst_duration,
-    const double& prf,
-    const double& processing_bandwidth  // B_d
-) {
-    int num_az = range_compressed.size();
-    int num_rng = range_compressed[0].size();
-
-    range_compressed = transpose(range_compressed);
-
-    compute_axis_dft_in_place(range_compressed, 0, 1, false);
-    fftshift_in_place(range_compressed);
-
-    double num_replicas = std::abs(dc_rate * burst_duration / prf);
-    double bandwidth = num_replicas * prf / 2.0;  // B_s
-
-    int shape = std::floor(num_replicas * num_az);
-    int mid = shape / 2;
-    int start = mid - (std::floor(processing_bandwidth) / 2);
-    int end = mid + (std::floor(processing_bandwidth) / 2);
-
-    std::cout << "-------------------------------------------------" << std::endl;
-    std::cout << "Computing Azimuth Frequency UFR with Parameters: " << std::endl;
-    std::cout << "PRF: " << prf << std::endl;
-    std::cout << "DC Rate: " << dc_rate << std::endl;
-    std::cout << "Burst Duration: " << burst_duration << std::endl;
-    std::cout << "Number of Spectral Replicas: " << num_replicas << std::endl;
-    std::cout << "Ramp Signal Bandwidth: " << bandwidth << std::endl;
-    std::cout << "Output Shape: " << shape << std::endl;
-    std::cout << "Low Pass Filter Center: " << mid << std::endl;
-    std::cout << "Low Pass Filter Start Index: " << start << std::endl;
-    std::cout << "Low Pass Filter End Index: " << end << std::endl;
-    std::cout << "-------------------------------------------------" << std::endl;
-
-    F_VEC_1D az_freqs = linspace(-bandwidth, bandwidth, shape);
-
-    CF_VEC_2D ufr_output(num_rng, CF_VEC_1D(shape));
-
-    #pragma omp parallel for
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        // Mosaic Signal
-        ufr_output[rng_line] = get_tiled_signal(range_compressed[rng_line], num_replicas);
-
-        //Apply De-ramping
-        for (int az_line = 0; az_line < shape; az_line++)
-        {
-            std::complex<double> sample = ufr_output[rng_line][az_line];
-            std::complex<double> deramp = std::exp(-I * PI * (1 / dc_rate) * std::pow(az_freqs[az_line], 2.0));
-
-            ufr_output[rng_line][az_line] = deramp * sample;
-        }
-    }
-
-    // Low-pass Filter
-    compute_axis_dft_in_place(ufr_output, 0, 1, false);
-
-    #pragma omp parallel for collapse(2)
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        for (int az_line = 0; az_line < shape; az_line++)
-        {
-            if (az_line < end-start)
-            {
-                ufr_output[rng_line][az_line] = ufr_output[rng_line][start + az_line];
-            }
-            else
-            {
-                ufr_output[rng_line][az_line] = 0.0;
-            }
-        }
-    }
-
-    compute_axis_dft_in_place(ufr_output, 0, 1, true);
-
-    #pragma omp parallel for collapse(2)
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        // Apply Re-ramping
-        for (int az_line = 0; az_line < shape; az_line++)
-        {
-            std::complex<double> sample = ufr_output[rng_line][az_line];
-            std::complex<double> reramp = std::exp(I * PI * (1 / dc_rate) * std::pow(az_freqs[az_line], 2.0));
-
-            ufr_output[rng_line][az_line] = reramp * sample;
-        }
-    }
-
-    return transpose(ufr_output);
-}
-
-
-CF_VEC_2D azimuth_time_ufr(
-    CF_VEC_2D& azimuth_compressed,
-    F_VEC_1D&  dc_estimates,
-    F_VEC_2D& ka,
-    L0Packet&  initial_packet,
-    const double& dc_rate,
-    const double& burst_duration,
-    const double& prf,
-    const double& processing_bandwidth  // B_d
-) {
-    ka = transpose(ka);
-    azimuth_compressed = transpose(azimuth_compressed);
-
-    int num_az = azimuth_compressed[0].size();
-    int num_rng = azimuth_compressed.size();
-
-    double num_replicas = std::abs(dc_rate * burst_duration / prf);
-    double bandwidth = num_replicas * prf;
-
-    double ka_mean = std::abs(std::accumulate(ka[0].begin(), ka[0].end(), 0.0)) / num_az;
-    double focused_time = burst_duration + ((bandwidth - (2 * processing_bandwidth)) / ka_mean);
-
-    int num_tiles = std::ceil((focused_time / 2) / burst_duration) - std::floor((-focused_time / 2) / burst_duration);
-
-    int output_shape = std::floor(prf * burst_duration);
-    int downsample_shape = output_shape + std::floor(processing_bandwidth);
-    int shape = num_tiles * num_az;
-    int mid = shape / 2;
-    int start = mid - bandwidth / 2;
-    int end = mid + bandwidth / 2;
-
-    std::cout << "-------------------------------------------------" << std::endl;
-    std::cout << "Computing Azimuth Time UFR with Parameters: " << std::endl;
-    std::cout << "PRF: " << prf << std::endl;
-    std::cout << "DC Rate: " << dc_rate << std::endl;
-    std::cout << "Burst Duration: " << burst_duration << std::endl;
-    std::cout << "Doppler Bandwidth: " << processing_bandwidth << std::endl;
-    std::cout << "Ramp Signal Bandwidth: " << bandwidth << std::endl;
-    std::cout << "Focused Time: " << focused_time << std::endl;
-    std::cout << "Mean Azimuth FM Rate: " << ka_mean << std::endl;
-    std::cout << "Number of Freq Spectral Replicas: " << num_replicas << std::endl;
-    std::cout << "Number of Time Spectral Replicas: " << num_tiles << std::endl;
-    std::cout << "Tiled Shape: " << shape << std::endl;
-    std::cout << "Low Pass Filter Center: " << mid << std::endl;
-    std::cout << "Low Pass Filter Start Index: " << start << std::endl;
-    std::cout << "Low Pass Filter End Index: " << end << std::endl;
-    std::cout << "Downsample Shape: " << downsample_shape << std::endl;
-    std::cout << "Output Shape: " << output_shape << std::endl;
-    std::cout << "-------------------------------------------------" << std::endl;
-
-    F_VEC_1D az_times = linspace(-focused_time / 2, focused_time / 2, shape);
-
-    CF_VEC_2D ufr_intermediate(num_rng, CF_VEC_1D(shape));
-
-    #pragma omp parallel for
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        // Mosaic Signal
-        ufr_intermediate[rng_line] = get_tiled_signal(azimuth_compressed[rng_line], num_tiles);
-
-        //Apply De-ramping
-        double f_dc = dc_estimates[rng_line];
-        F_VEC_1D ka_row = linear_resample(ka[rng_line], shape);
-
-        for (int az_line = 0; az_line < shape; az_line++)
-        {
-            double kt = -ka_row[az_line] * dc_rate / (dc_rate - ka_row[az_line]);
-            double dc_time = -f_dc / ka_row[az_line];
-            double az_time = dc_time + az_times[az_line];
-
-            std::complex<double> deramp = std::exp( (-1.0 * I * PI * kt * std::pow(az_time, 2.0)) + (2.0 * I * PI * (kt + ka_row[az_line]) * dc_time * az_time) );
-            std::complex<double> sample = ufr_intermediate[rng_line][az_line];
-
-            ufr_intermediate[rng_line][az_line] = deramp * sample;
-        }
-    }
-
-    // Low-pass Filter
-    compute_axis_dft_in_place(ufr_intermediate, 0, 1, false);
-
-    #pragma omp parallel for collapse(2)
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        for (int az_line = 0; az_line < shape; az_line++)
-        {
-            if (az_line < end-start)
-            {
-                ufr_intermediate[rng_line][az_line] = ufr_intermediate[rng_line][start + az_line];
-            }
-            else
-            {
-                ufr_intermediate[rng_line][az_line] = 0.0;
-            }
-        }
-    }
-
-    compute_axis_dft_in_place(ufr_intermediate, 0, 1, true);
-
-    // Resample
-    CF_VEC_2D ufr_output(num_rng, CF_VEC_1D(downsample_shape));
-
-    std::transform(
-        ufr_intermediate.begin(), ufr_intermediate.end(),
-            ufr_output.begin(),
-                [downsample_shape] (CF_VEC_1D& ufr_row) {
-                    return linear_resample(ufr_row, downsample_shape);
-                }
-    );
-
-    ufr_intermediate.clear();
-    ufr_intermediate.shrink_to_fit();
-
-    // Apply Re-ramping
-    #pragma omp parallel for
-    for (int rng_line = 0; rng_line < num_rng; rng_line++)
-    {
-        double f_dc = dc_estimates[rng_line];
-        F_VEC_1D ka_row = linear_resample(ka[rng_line], downsample_shape);
-
-        for (int az_line = 0; az_line < downsample_shape; az_line++)
-        {
-            double kt = -ka_row[az_line] * dc_rate / (dc_rate - ka_row[az_line]);
-            double dc_time = -f_dc / ka_row[az_line];
-            double az_time = dc_time + az_times[az_line];
-
-            std::complex<double> reramp = std::exp( (1.0 * I * PI * kt * std::pow(az_time, 2.0)) - (2.0 * I * PI * (kt + ka_row[az_line]) * dc_time * az_time) );
-            std::complex<double> sample = ufr_output[rng_line][az_line];
-
-            ufr_output[rng_line][az_line] = reramp * sample;
-        }
-    }
-
-    ufr_output = transpose(ufr_output);
-
-    // Crop to Valid Portion
-    mid = downsample_shape / 2;
-    start = mid - (output_shape / 2);
-    end = mid + (output_shape / 2);
-
-    return CF_VEC_2D(ufr_output.begin() + start, ufr_output.begin() + end);
-}
