+++ src/image_write.cpp
@@ -69,31 +69,6 @@ void write_swath(
 }
 
 
-void write_burst_replica_chirps(
-    const std::string& in_filename,
-    const std::string& out_filename,
-    const std::string& swath_name,
-    const int&         burst_num,
-    const std::string& scaling_mode
-) {
-    Burst burst(in_filename, swath_name, burst_num);
-    CF_VEC_2D chirps = burst.get_replica_chirps();
-    write_tif(chirps, out_filename, scaling_mode);
-}
-
-
-void write_swath_replica_chirps(
-    const std::string& in_filename,
-    const std::string& out_filename,
-    const std::string& swath_name,
-    const std::string& scaling_mode
-) {
-    Swath swath(in_filename, swath_name);
-    CF_VEC_2D chirps = swath.get_all_replica_chirps();
-    write_tif(chirps, out_filename, scaling_mode);
-}
-
-
 void write_range_compressed_swath(
     const std::string& in_filename,
     const std::string& out_filename,
