+++ src/s1_decoder.cpp
@@ -148,32 +148,6 @@ CF_VEC_2D S1_Decoder::get_swath(const std::string& swath)
 }
 
 
-std::pair<PACKET_VEC_2D, int> S1_Decoder::get_azimuth_blocks(PACKET_VEC_1D& packets)
-{
-    PACKET_VEC_2D azimuth_blocks;
-    PACKET_VEC_1D azimuth_block;
-    
-    int previous_size = 2 * packets[0].get_num_quads();
-    int max_size = previous_size;
-
-    for (int i = 0; i < packets.size(); i++)
-    {
-        L0Packet packet = packets[i];
-        int size = 2 * packet.get_num_quads();
-        if (size != previous_size or i == packets.size() - 1)
-        {
-            if (size > max_size) max_size = size;
-            previous_size = size;
-            azimuth_blocks.push_back(azimuth_block);
-            azimuth_block = {};
-        }
-        azimuth_block.push_back(packet);
-    }
-
-    return std::pair<PACKET_VEC_2D, int>(azimuth_blocks, max_size);
-}
-
-
 CF_VEC_2D S1_Decoder::get_range_compressed_burst(const std::string& swath, const int& burst, bool range_doppler)
 {
     PACKET_VEC_1D burst_packets = _echo_packets[swath][burst];
